+++ tests/test_version.py
@@ -2,4 +2,4 @@ from alvis import __version__
 
 
 def test_version() -> None:
-    assert __version__ == "0.6.0"
+    assert __version__ == "1.0.0rc2"